refactor: modularize ChatWidget into separate components
